requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
email-validator==2.2.0

motor==3.6.0
pymongo==4.9.2

argon2-cffi==23.1.0
twilio==9.0.4
PyJWT==2.9.0

tensorflow==2.17.1
joblib==1.4.2
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.7.2
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.63
langsmith==0.1.147
ollama==0.3.3

googlesearch-python==1.2.5
beautifulsoup4==4.12.3
aiohttp==3.11.10
requests==2.32.3


